Straddling Cells has correct answer
Thought about this better. We can increment over cells in a more logical
manner.
*  *
^ /
|/
*->*
We only need to look at these points and compare them to the lower left
corner. If we do divisions, we only need to check if they are negative
(assuming that the lower left value isn't 0).