Split Window (read about issue)
Split windows is implemented. Euler on left and RK4 on the right.
Issues:
RK4 does not seem to outperform Euler. In fact, the top right streamline
changes. The Euler value stays the same but the RK4 diverges in answer
when h is increased.
With h = 0.01 and nsteps = 5000 answers match
With h=0.5 and nsteps=400 answers diverge, Euler looks similar to
original